MCP_Folder/github_pr_reviewer/requirements.txt
fastapi
uvicorn
fastmcp
httpx
python-dotenv
pydantic
loguru
groq